Administrator user must be protected and can't be deleted

+++ module/Config/src/Config/Controller/UserController.php
@@ -32,6 +32,7 @@ use Config\Form\UserLogin,
     Config\Form\UserForgotPassword as UserForgotForm,
     Gc\Mvc\Controller\Action,
     Gc\User,
+    Gc\User\Role,
     Zend\Http\Request,
     Zend\View\Model\ViewModel,
     DateTime;
@@ -60,8 +61,16 @@ class UserController extends Action
     public function indexAction()
     {
         $user_collection = new User\Collection();
+        $users = array();
+        foreach($user_collection->getUsers() as $user)
+        {
+            if($user->getRole()->getName() !== Role\Model::PROTECTED_NAME)
+            {
+                $users[] = $user;
+            }
+        }
 
-        return array('users' => $user_collection->getUsers());
+        return array('users' => $users);
     }
 
     /**
@@ -215,7 +224,7 @@ class UserController extends Action
     public function deleteAction()
     {
         $user = User\Model::fromId($this->getRouteMatch()->getParam('id'));
-        if(empty($user))
+        if(empty($user) and $user->getRole()->getName() !== Role\Model::PROTECTED_NAME)
         {
             if($user->delete())
             {
